status.c: use a signed comparison when clamping the field width
The field-width clamp compared an int against sizeof (unsigned), so a
negative width promoted to a large unsigned value and skipped the clamp.
Cast the size to int to keep the comparison signed.

+++ source/status.c
@@ -1412,7 +1412,7 @@ status_query_nick(window)
                 int buflen = get_int_var(CHANNEL_NAME_WIDTH_VAR);
                 char tmpbuf[mybufsize / 4];
 
-                if (buflen > sizeof(tmpbuf)) buflen = sizeof(tmpbuf);
+                if (buflen > (int)sizeof(tmpbuf)) buflen = (int)sizeof(tmpbuf);
                 strmcpy(tmpbuf, window->query_nick, buflen);
 		snprintf(lbuf, sizeof(lbuf), query_format, tmpbuf);
 /****************************************************************************/
